moving stuff round
