--- src/game/actors/heros/barbarian.ts
@@ -1,5 +1,0 @@
-import TGHero from "../hero/hero";
-
-export default class TGBarbarian extends TGHero {
-
-}
