+++ src/game/main.ts
@@ -1,13 +1,10 @@
 import HPApp from '../engine/core/app';
 import HPVector from '../engine/physics/vector';
 import TGActorFactory from './actor-factory';
-import TGWanderingTarget from './actors/wandering-target';
-import TGWizard from './actors/heros/wizard';
 import TGHero from './actors/hero/hero';
 
 const assets = [
   TGHero.textureFile,
-  TGWanderingTarget.imageFile,
 ];
 
 const app = new HPApp({
@@ -15,7 +12,7 @@ const app = new HPApp({
   actorFactory: TGActorFactory,
   areaFile: 'public/areas/test-1.json',
   assets,
-  hero: new TGWizard(),
+  hero: new TGHero(),
   heroStart: new HPVector(200, 700),
 });
